Main/requirements.txt
   streamlit
    Pillow
    python-dotenv
    requests
    openai
    chromadb
    sentence-transformers